* Changing tag status to int8_t and reorganized tag elements to reduce struct size.

* Adding test program for auto sync.

* Adding test program for auto sync.

* Update version number.

* Set the read_in_flight flag when kicking off initial read.

* Add first cut of automatic read/write.

* Fix commented out mutex_lock/unlock.

* Missed a commit of code to clear the in flight flags.

* Fixed debug level.

* Removed debug level.

* More clean up around read/write_in_flight flags.

* Changing tag status to int8_t and reorganized tag elements to reduce struct size.

* Adding test program for auto sync.

* Adding test program for auto sync.

* Set the read_in_flight flag when kicking off initial read.

* Add first cut of automatic read/write.

* Fix commented out mutex_lock/unlock.

* Missed a commit of code to clear the in flight flags.

* Fixed debug level.

* Removed debug level.

* More clean up around read/write_in_flight flags.

* Changing tag status to int8_t and reorganized tag elements to reduce struct size.

* Adding test program for auto sync.

* Adding test program for auto sync.

* Set the read_in_flight flag when kicking off initial read.

* Add first cut of automatic read/write.

* Fix commented out mutex_lock/unlock.

* Missed a commit of code to clear the in flight flags.

* Fixed debug level.

* Removed debug level.

* More clean up around read/write_in_flight flags.

* Add first cut of automatic read/write.

* More clean up around read/write_in_flight flags.

* Added indentation for clarity.

* Fixed up timing to make automatic reads try hard to be exactly on schedule.

* Fixed typo in tag attribute string.

* Update version to 2.1.12 for auto tags.

* Need to rewrite test_auto_sync for Windows.

+++ src/protocols/ab/eip_cip.c
@@ -200,7 +200,7 @@ int tag_tickler(ab_tag_p tag)
             rc = check_read_status_unconnected(tag);
         }
 
-        tag->status = rc;
+        tag->status = (int8_t)rc;
 
         /* if the operation completed, make a note so that the callback will be called. */
         if(!tag->read_in_progress) {
@@ -219,7 +219,7 @@ int tag_tickler(ab_tag_p tag)
             rc = check_write_status_unconnected(tag);
         }
 
-        tag->status = rc;
+        tag->status = (int8_t)rc;
 
         /* if the operation completed, make a note so that the callback will be called. */
         if(!tag->write_in_progress) {
